// src/app/model/contact.ts
export class Contact {
  ContactId: any
  IdentityNumber: string
  FirstName: string
  LastName: string
  BirthdayDate: Date
  GenderId: any
  Address: string
  CityId: any
  CityName: string
  Phone: string
  Phone2: string
  IsSmsAllowed: boolean
  Email: string
  IsEmailAllowed: boolean
  HashalNumber: string
  HashalRankId: any
  HashalRankName: string
  HashalLastJob: string
  HashalUnit: string
  IsHashalPension: boolean
  DraftDate?: Date
  DischargeDate?: Date
  MaritalStatus: string
  CurrentCompany: string
  CurrentJob: string
  Phone3: string
  CurrentJobAddress: string
  HashalDonateAmount: number
  IsMonthlyChargeApproval?: boolean
  MonthlyChargeApprovalDate?: Date
  CreateUser: string
  ContactCreator: string
  CreateDate: Date
  UpdateUser: string
  ContactUpdater: string
  UpdateDate?: Date
  HashalServiceTypeId: any

  Token: string
  TransactionId: string
  CurrencyCode: string
  CardLast4Digits: string
  ValidityMonth: any
  ValidityYear: any
  HowManyPayments: any
  CardHolderIdNumber: string
};